added plot fun for output file with best_sol

+++ input.c
@@ -75,7 +75,7 @@ void read_input(instance* inst)
             inst->xcoord = (double*)calloc(inst->nnodes, sizeof(double));
             inst->ycoord = (double*)calloc(inst->nnodes, sizeof(double));
             inst->index=(int*)calloc(inst->nnodes, sizeof(int));
-            inst->cost=(double*)calloc(inst->nnodes*inst->nnodes,sizeof(int));
+            inst->cost=(double*)calloc(inst->nnodes*inst->nnodes,sizeof(double));
             inst->flagCost = 0;
             active_section = 0;
             continue;
